src/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy<2.0
chromadb==0.4.24
sentence-transformers==2.2.2
torch>=1.13.0
transformers>=4.21.0
google-generativeai>=0.3.0